+++ packages/sdk/src/adaptors/index.ts
@@ -24,6 +24,7 @@ export { default as getGovernordaoData } from './governordao';
 export { default as getHiveOneData } from './hiveone';
 export { default as checkIdena } from './idena';
 export { default as getKarmaData } from './karma';
+export { default as getKlimaData } from './klima';
 export { default as getKnownOriginData } from './knownorigin';
 export { default as getLayer3Data } from './layer3';
 export { default as getLensData } from './lens';